refactor home screen UI
